move currentuserslice to user folder
