+++ src/redux/configStore.js
@@ -7,7 +7,7 @@ import loginSlice from './auth/loginSlice';
 import logoutSlice from './auth/logoutSlice';
 import addHouseSlice from './house/addHouseSlice';
 import displayHouseSlice from './house/displayHouseSlice';
-import currentUserSlice from './house/currentUserSlice';
+import currentUserSlice from './user/currentUserSlice';
 
 // root Reducer
 const rootReducer = combineReducers({
